+++ EDDI/EDDI.cs
@@ -783,10 +783,6 @@ namespace EddiCore
                     {
                         passEvent = eventCarrierJumped((CarrierJumpedEvent)@event);
                     }
-                    else if (@event is SignalDetectedEvent)
-                    {
-                        passEvent = eventSignalDetected((SignalDetectedEvent)@event);
-                    }
 
                     // Additional processing is over, send to the event responders if required
                     if (passEvent)
@@ -810,16 +806,6 @@ namespace EddiCore
                     Instance.ObtainResponder("EDDN responder").Handle(@event);
                 }
             }
-        }
-
-        private bool eventSignalDetected(SignalDetectedEvent @event) 
-        {
-            if (Instance.CurrentStarSystem != null && Instance.CurrentStarSystem.systemAddress == @event.systemAddress)
-            {
-                Instance.CurrentStarSystem.AddOrUpdateSignalSource(@event.signalSource);
-                return true;
-            }
-            return false;
         }
 
         private bool eventCarrierJumpEngaged(CarrierJumpEngagedEvent @event)